requirements.txt
alabaster==0.7.12
anaconda-client==1.7.2
anaconda-navigator==1.9.7
anaconda-project==0.8.3
argh==0.26.2
asn1crypto==1.2.0
astroid==2.3.2
astropy==4.0
atomicwrites==1.3.0
attrs==19.3.0
autopep8==1.4.4
Babel==2.7.0
backcall==0.1.0
backports.functools-lru-cache==1.6.1
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==3.1.7
beautifulsoup4==4.8.1
bitarray==1.2.0
bkcharts==0.2
bleach==3.1.0
bokeh==1.4.0
boto==2.49.0
Bottleneck==1.3.1
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
clyent==1.2.2
colorama==0.4.1
comtypes==1.1.7
conda==4.8.1
conda-build==3.18.8
conda-package-handling==1.6.0
conda-verify==3.4.2
contextlib2==0.6.0.post1
cryptography==2.8
cycler==0.10.0
Cython==0.29.14
cytoolz==0.10.1
dask==2.9.1
decorator==4.4.1
defusedxml==0.6.0
diff-match-patch==20181111
distributed==2.9.1
dj-database-url==0.5.0
Django==1.11.17
django-countries==5.5
django-heroku==0.3.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
docutils==0.15.2
entrypoints==0.3
et-xmlfile==1.0.1
Faker==2.0.3
fastcache==1.1.0
filelock==3.0.12
flake8==3.7.9
Flask==1.1.1
fsspec==0.6.2
future==0.18.2
gevent==1.4.0
glob2==0.7
greenlet==0.4.15
gspread==3.1.0
gunicorn==20.0.4
h5py==2.9.0
HeapDict==1.0.1
html5lib==1.0.1
httplib2==0.14.0
hypothesis==4.54.2
idna==2.8
imageio==2.6.1
imagesize==1.1.0
importlib-metadata==1.3.0
intervaltree==3.0.2
ipykernel==5.1.3
ipython==7.11.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.14.1
Jinja2==2.10.3
joblib==0.14.1
json5==0.8.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==5.2.0
jupyter-core==4.6.1
jupyterlab==1.2.4
jupyterlab-server==1.0.6
keyring==20.0.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.2
libarchive-c==2.8
llvmlite==0.29.0
locket==0.2.0
lxml==4.4.2
MapProxy==1.12.0
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
menuinst==1.4.16
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
mock==3.0.5
more-itertools==8.0.2
mpmath==1.1.0
msgpack==0.6.1
multipledispatch==0.6.0
navigator-updater==0.2.1
nbconvert==5.6.1
nbformat==4.4.0
networkx==2.4
nltk==3.4.5
nose==1.3.7
notebook==6.0.2
numba==0.44.1
numexpr==2.7.0
numpy==1.17.4
numpydoc==0.9.1
oauth2client==4.1.3
olefile==0.46
openpyxl==3.0.2
packaging==19.2
pandas==0.25.3
pandocfilters==1.4.2
paramiko==2.6.0
parso==0.5.2
partd==1.1.0
path==13.1.0
pathlib2==2.3.5
pathtools==0.1.2
patsy==0.5.1
pep8==1.7.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
pipenv==2018.11.26
pkginfo==1.5.0.1
pluggy==0.13.1
ply==3.11
prometheus-client==0.7.1
prompt-toolkit==3.0.2
psutil==5.6.7
psycopg2==2.8.4
py==1.8.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.3
pydocstyle==4.0.1
pyflakes==2.1.1
Pygments==2.5.2
pylint==2.4.3
Pympler==0.7
PyNaCl==1.3.0
pyodbc==4.0.27
pyOpenSSL==19.1.0
pyparsing==2.4.5
pyproj==2.4.1
pyreadline==2.1
pyrsistent==0.15.6
pyshp==2.1.0
PySocks==1.7.1
pytest==5.3.2
pytest-arraydiff==0.3
pytest-astropy==0.7.0
pytest-astropy-header==0.1.1
pytest-doctestplus==0.5.0
pytest-openfiles==0.4.0
pytest-remotedata==0.3.2
python-dateutil==2.8.1
python-jsonrpc-server==0.3.2
python-language-server==0.31.2
pytz==2019.3
PyWavelets==1.1.1
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.2
pyzmq==18.1.0
QDarkStyle==2.7
qrcode==6.1
QtAwesome==0.6.0
qtconsole==4.6.0
QtPy==1.9.0
requests==2.22.0
rope==0.14.0
rsa==4.0
Rtree==0.8.3
ruamel-yaml==0.15.87
scikit-image==0.15.0
scikit-learn==0.22
scipy==1.3.2
seaborn==0.9.0
selenium==3.141.0
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.12.0
snowballstemmer==2.0.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.9.5
Sphinx==2.3.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.1.2
spyder==4.0.0
spyder-kernels==1.8.1
SQLAlchemy==1.3.11
statsmodels==0.10.1
sympy==1.5
tables==3.6.1
tblib==1.6.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
toolz==0.10.0
tornado==6.0.3
tqdm==4.40.2
traitlets==4.3.3
typed-ast==1.4.0
ujson==1.35
unicodecsv==0.14.1
urllib3==1.25.7
virtualenv==16.7.9
virtualenv-clone==0.5.3
watchdog==0.9.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
whitenoise==5.0.1
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.11.2
xlrd==1.2.0
XlsxWriter==1.2.6
xlwings==0.16.3
xlwt==1.3.0
yapf==0.28.0
zict==1.0.0
zipp==0.6.0
